fix: optimized aggregation call hierarchy and prevent unnecessary calls
